mainly fixed some graph things, polished many other things

+++ src/qpl/perlin_noise.cpp
@@ -9,9 +9,10 @@ namespace qpl {
 		this->progress_x += fx;
 		this->progress_y += fy;
 		this->value = this->noise(this->progress_x, this->progress_y, this->frequency, this->octaves);
+
 		return this->value;
 	}
-	void qpl::perlin_noise_generator::set_octaves(qpl::f64 octaves) {
+	void qpl::perlin_noise_generator::set_octaves(qpl::u32 octaves) {
 		this->octaves = octaves;
 	}
 	void qpl::perlin_noise_generator::set_frequency(qpl::f64 frequency) {
